Simplify colors

+++ Source/Pitch.cpp
@@ -4,13 +4,14 @@
 
 #include "Pitch.h"
 #include "PitchClass.h"
+#include "Hash.h"
 
-Pitch::Pitch(float midiPitch)
+Pitch::Pitch(double midiPitch)
 {
 	this->midiPitch = midiPitch;
 }
 
-Pitch Pitch::fromFreqHz(float freqHz)
+Pitch Pitch::fromFreqHz(double freqHz)
 {
 	return Pitch(69.f + (log2f(freqHz) - log2f(440)) * 12.f);
 }
@@ -31,7 +32,7 @@ bool Pitch::operator<(const Pitch& pitch) const
 	return midiPitch < pitch.midiPitch;
 }
 
-float Pitch::getMidiPitch() const 
+double Pitch::getMidiPitch() const 
 {
 	return this->midiPitch;
 }
