+++ Source/Pitch.h
@@ -8,24 +8,24 @@ class PitchClass;
 class Pitch
 {
 public:
-	Pitch(float);
+	Pitch(double);
 	Pitch(const Pitch&);
-	static Pitch fromFreqHz(float);
+	static Pitch fromFreqHz(double);
 	bool operator==(const Pitch&) const;
 	bool operator<(const Pitch&) const;
-	float getMidiPitch() const;
+	double getMidiPitch() const;
 
 	// Two Pitches are equal if their difference is below this value
 	// 1/1000 cent
-	static constexpr float epsilon = 0.00001; 
+	static constexpr double epsilon = 0.00001; 
 private: 
-	float midiPitch; 
+	double midiPitch; 
 };
 
 struct PitchHash {
 public:
 	size_t operator()(const Pitch& x) const
 	{
-		return std::hash<float>()((int)x.getMidiPitch());
+		return std::hash<double>()((int)x.getMidiPitch());
 	}
 };
